refactor: simplify accessTokenParser import paths across routes
